+++ analyzer/src/main/java/nl/basjes/parse/useragent/AbstractUserAgentAnalyzerDirect.java
@@ -216,6 +216,7 @@ public abstract class AbstractUserAgentAnalyzerDirect implements Analyzer, Analy
             // Lastly we modify the parsed userAgent with the found clientHints
             userAgent = clientHintsAnalyzer.merge(userAgent, clientHints);
         }
+        userAgent = matchMaker.hardCodedPostProcessing(userAgent);
         return new ImmutableUserAgent(userAgent);
     }